pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "graphrag-weaviate"
version = "0.1.0"
description = "Internal pipeline for validating Russian clinic appointment data against MKB-scoped clinical recommendations"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "weaviate-client>=4.10.0",
  "langchain-ollama>=0.2.0",
  "pymupdf>=1.24.0",
  "python-dotenv>=1.0.1",
  "pydantic>=2.8.0",
]

[tool.setuptools]
packages = ["graphrag_weaviate"]
